added some fixes for linux / mac

+++ renderer.cpp
@@ -4,6 +4,7 @@
 
 #include <cstdlib>
 #include <cstddef>
+#include <cstdint>
 
 namespace Sndspec {
 
@@ -27,14 +28,14 @@ void Renderer::Render(const Parameters &parameters, const SpectrogramResults<dou
 	int numBins = spectrogramData.at(0).at(0).size();
 	int h = height - 1;
 	double colorScale = heatMapPalette.size() / -parameters.getDynRange();
-	int lastColorIndex = std::max(0ULL, heatMapPalette.size() - 1);
+	unsigned long lastColorIndex = std::max(0ul, heatMapPalette.size() - 1);
 
 	for(int c = 0; c < 1 /*numChannels*/; c++) {
 		for(int x = 0; x < numSpectrums; x++) {
 			for(int y = 0; y < numBins; y++) {
 				int addr = x + (h - y) * stride32;
-				int colorindex = spectrogramData[c][x][y] * colorScale;
-				int32_t color = heatMapPalette[std::max(0, std::min(colorindex, lastColorIndex))];
+				unsigned long colorindex = spectrogramData[c][x][y] * colorScale;
+				int32_t color = heatMapPalette[std::max(0ul, std::min(colorindex, lastColorIndex))];
 				pixelBuffer[addr] = color;
 			}
 		}
